Подключит слайсы в Navigation
